Proper {at,de}tach / {ac,deac}tivate split
Previously the mapping of SMAPI commands to tapdisk actions was:
 * attach => tap.create; tap.open
 * activate => noop
 * deactivate => tap.close; tap.destroy
 * detach => noop

Now the mapping of SMAPI commands to tapdisk actions is:
 * attach => tap.create
 * activate => tap.open
 * deactivate => tap.close
 * detach => tap.destroy

To achieve this, the plugin must remember the tapdisk created in attach so that
it can be referred to in activate, deactivate and detach. (Note that it is not
possible to rely on tapdisk.find_by_file as the filename is only associated with
the tapdisk instance after activate and before detach.) This is achieved by
pickling the Tapdisk object to non-persistent storage on the local host and
unpickling it when required.

Although it incurs filesystem I/O, it avoids the need to call
tapdisk.find_by_file -- which invokes tap-ctl list, which can sometimes be very
slow -- so is likely to improve performance, although this has not been
measured.

(Note that it is an explicit design decision for SMAPIv3 datapath plugins to
manage their own state to avoid making assumptions about other layers, so this
approach is preferable to relying on the caller to carry tapdisk metadata across
invocations of datapath plugin functions.)

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>